Added some properties to Dto interfaces. Changed AnonymousUserGUID form UUID to Guid Added high precision timer to Portal Requests, the duration of the call on the server is returned in the Portal Response Header

+++ src/app/Chaos.Portal/Response/IPortalHeader.cs
@@ -7,7 +7,7 @@ namespace Chaos.Portal.Response
     {
         [Serialize("Duration")]
         [SerializeXML(true)]
-        uint Duration { get; }
+        double Duration { get; }
 
         ReturnFormat ReturnFormat { get; set; }
     }
